pyproject.toml
[tool.poetry]
name = "telegram-bot-vercel"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9" # Укажите вашу версию Python
python-telegram-bot = "^20.7" # Актуальная версия библиотеки
flask = "^3.0.0" # Для создания веб-сервера под Vercel
python-dotenv = "^1.0.0" # Для загрузки .env

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"